requirement.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.2.0
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-heroku==0.3.1
django-phone-verify==2.0.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
nexmo==2.5.2
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
twilio==7.9.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.3.0
wrapt==1.14.1
